requirements.txt
# Dependencias principales
requests>=2.31.0
pandas>=2.0.0
tabulate>=0.9.0
python-dotenv>=1.0.0

# Interface web
streamlit>=1.28.0
plotly>=5.17.0
altair>=5.1.0

# Export to Excel
openpyxl>=3.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
responses>=0.23.0

# Development
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# Logging y utilidades
rich>=13.6.0